[AF022] No ActionBar onActivityResult
In AndroidMainfest.xml the theme is changed to @style/Theme.Design.NoActionBar to fit all new elements inside the main and second activity in the app.  (I removed the action bar)

In MainActivity addItemFromSecondAct is created where we start the second Activity with the right req_code for resulst and onActivityResult is implemented to get the right result.

In SecondActivity an onClick method is called and the text of every clicked button is put to an extra of an intent and the intent ist then set as Result and the activity is finished.