[AF022] Added items from second act to main tV´s
In second activty the user an now click on a button and the value of the button (the text label of it) will be displayed in a empty textview (if not all ten are filled)
Also when the rotation is changed or the main activity is newly created the itmes in the view would be saved in the savedInstanceState.
Chellange is finished.