[AF023] inplicite Intent for geo data
In Main Activity I created an Edit Text and a Button to get geo data from the input over a button click and implicite intent. To enable the intent to get resolve I needet a intent filter in the Manifest file with scheme geo for the action view.  An Uri then hase to be create with the right scheme plus the input and passed to the intent with the right action. Then the intent has to be resolved bevor startet with the activity and checked if the getPackageManger for it is not null.